Update admin.py (#3)
Added functionality to add/update/delete library branches and borrowers. Also, added functionality to override due date. The date validation could use a little cleaning up but it works fine if a valid date is entered. Also when it lists the book loans for a specific borrower, it just lists the loan table details, and isn't very pretty to look at.